AccountDialog JavaFX-ified phase 1 :+1:  :sunglasses: :egg:
Changed most of the JObject into the JavaFX objects and fixed what i
could.
Some of the object are specific to awt so it might be possible to use
them, I am going to ask him tomorrow about: Desktop, Transferable,
StringSelection, Container and GridbagConstraints. If these can be used
it will fix most of the errors within the class, the rest of the errors
would be the events